refactor: moved error types and phrases to package
