+++ src/clients/api/queries/getMarketHistory/index.ts
@@ -1,6 +1,7 @@
-import { VError } from 'errors';
 import { MarketSnapshot, VToken } from 'types';
 import { restService } from 'utilities';
+
+import { VError } from 'packages/errors/VError';
 
 const ENTRIES_PER_HOUR = 2;
 const ENTRIES_PER_DAY = ENTRIES_PER_HOUR * 24;
