+++ src/clients/api/queries/getPendingRewards/index.ts
@@ -1,7 +1,7 @@
 import BigNumber from 'bignumber.js';
 import { logError } from 'errors';
 
-import { VError } from 'errors/VError';
+import { VError } from 'packages/errors/VError';
 import convertPriceMantissaToDollars from 'utilities/convertPriceMantissaToDollars';
 import extractSettledPromiseValue from 'utilities/extractSettledPromiseValue';
 import findTokenByAddress from 'utilities/findTokenByAddress';
