+++ src/clients/api/queries/getProposals/index.ts
@@ -1,5 +1,6 @@
-import { VError } from 'errors';
 import { formatToProposal, restService } from 'utilities';
+
+import { VError } from 'packages/errors/VError';
 
 import { GetProposalsInput, GetProposalsOutput, ProposalsApiResponse } from './types';
