+++ src/clients/api/queries/getTransactions/index.ts
@@ -1,6 +1,7 @@
-import { VError } from 'errors';
 import { Token, Transaction, TransactionEvent, VToken } from 'types';
 import { restService } from 'utilities';
+
+import { VError } from 'packages/errors/VError';
 
 import formatTransaction from './formatTransaction';
 import { TransactionResponse } from './types';
