+++ src/clients/api/queries/getVoterAccounts/index.ts
@@ -1,7 +1,8 @@
 import BigNumber from 'bignumber.js';
-import { VError } from 'errors';
 import { VoterAccount } from 'types';
 import { restService } from 'utilities';
+
+import { VError } from 'packages/errors/VError';
 
 import formatVoterAccountResponse from './formatVoterAccountResponse';
 import { GetVoterAccountsResponse } from './types';
