+++ src/clients/api/queries/getVoterDetails/index.ts
@@ -1,6 +1,7 @@
-import { VError } from 'errors';
 import { Voter } from 'types';
 import { restService } from 'utilities';
+
+import { VError } from 'packages/errors/VError';
 
 import formatVoterDetailsResponse from './formatVoterDetailsResponse';
 import { GetVoterDetailsResponse } from './types';
