+++ src/clients/api/queries/getVoterHistory/index.ts
@@ -1,6 +1,7 @@
-import { VError } from 'errors';
 import { VoterHistory } from 'types';
 import { restService } from 'utilities';
+
+import { VError } from 'packages/errors/VError';
 
 import formatVoterHistoryResponse from './formatVoterHistoryResponse';
 import { GetVoterHistoryResponse } from './types';
