+++ src/errors/index.ts
@@ -1,4 +1,2 @@
-export * from './VError';
-export * from './transactionErrors';
 export * from './logError';
 export * from './displayMutationError';
