+++ src/hooks/useOperationModal/Modal/RepayForm/index.tsx
@@ -9,7 +9,6 @@ import {
   SwapDetails,
   TokenTextField,
 } from 'components';
-import { VError } from 'errors';
 import { useGetSwapRouterContractAddress } from 'packages/contracts';
 import { useTranslation } from 'packages/translations';
 import React, { useCallback, useMemo, useState } from 'react';
@@ -29,6 +28,7 @@ import useGetSwapInfo from 'hooks/useGetSwapInfo';
 import useGetSwapTokenUserBalances from 'hooks/useGetSwapTokenUserBalances';
 import { useIsFeatureEnabled } from 'hooks/useIsFeatureEnabled';
 import useTokenApproval from 'hooks/useTokenApproval';
+import { VError } from 'packages/errors/VError';
 
 import { useStyles as useSharedStyles } from '../styles';
 import Notice from './Notice';
