+++ src/packages/errors/VError.ts
