+++ src/packages/wallet/hooks/useLogIn/index.ts
@@ -1,5 +1,5 @@
 import { openInfinityWallet } from '@infinitywallet/infinity-connector';
-import { VError } from 'errors';
+import { VError } from 'packages/errors';
 import { connectorIdByName } from 'packages/wallet/connectors';
 import { Connector } from 'packages/wallet/types';
 import { useCallback } from 'react';
