+++ src/pages/Vai/MintVai/index.tsx
@@ -7,7 +7,7 @@ import {
   LabeledInlineContent,
   Spinner,
 } from 'components';
-import { VError, displayMutationError } from 'errors';
+import { displayMutationError } from 'errors';
 import { useGetToken } from 'packages/tokens';
 import { useTranslation } from 'packages/translations';
 import React, { useCallback, useMemo } from 'react';
@@ -31,6 +31,7 @@ import { AmountForm } from 'containers/AmountForm';
 import { ConnectWallet } from 'containers/ConnectWallet';
 import { useAuth } from 'context/AuthContext';
 import useConvertMantissaToReadableTokenString from 'hooks/useConvertMantissaToReadableTokenString';
+import { VError } from 'packages/errors/VError';
 
 import { useStyles } from '../styles';
 import getReadableFeeVai from './getReadableFeeVai';
